# openapi.yaml
openapi: 3.0.0
info:
  title: RFQ Quote Management API
  version: 1.1.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [admin, buyer, vendor] }
        name: { type: string }
        email: { type: string, format: email }
        slug: { type: string }
        company_name: { type: string }
        bio: { type: string }
        phone: { type: string }
        website: { type: string }
        address: { type: object }
        business_info: { type: object }
        rating: { type: number }
...
    RFQ:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        slug_title: { type: string }
        description: { type: string }
        buyer_id: { type: string, format: uuid }
        category_id: { type: integer }
        subcategory_id: { type: integer }
        tags: { type: array, items: { type: string } }
        budget_min: { type: number }
        budget_max: { type: number }
        currency: { type: string }
        status: { type: string, enum: [open, closed, awarded, cancelled] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NewRFQ:
      type: object
      required: [ title, description, category_id ]
      properties:
        title: { type: string }
        description: { type: string }
        category_id: { type: integer }
        subcategory_id: { type: integer }
        tags: { type: array, items: { type: string } }
        budget_min: { type: number }
        budget_max: { type: number }
        currency: { type: string }
...
paths:
  /auth/otp/request:
    post:
      summary: Request an OTP to be sent by email
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                purpose: { type: string, enum: [registration, login, reset_password], default: registration }
      responses:
        '201':
          description: OTP created
  /auth/otp/verify:
    post:
      summary: Verify OTP and obtain JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string }
                purpose: { type: string, enum: [registration, login, reset_password], default: registration }
                name: { type: string }
                role: { type: string, enum: [buyer, vendor] }
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
...
  /rfq:
    get:
      summary: Browse and filter RFQs
      security: [{ bearerAuth: [] }]
      tags: [RFQ]
      parameters:
        - in: query
          name: category_id
          schema: { type: integer }
        - in: query
          name: subcategory_id
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: tags
          schema: { type: array, items: { type: string } }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List RFQs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RFQ' }
    post:
      summary: Create a new RFQ (buyer only)
      security: [{ bearerAuth: [] }]
      tags: [RFQ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NewRFQ' }
      responses:
        '201': { description: RFQ created }
...